Slightly improved file structure
