+++ server/events/mod.ts
@@ -1,9 +1,34 @@
-export { default as handleJoin } from "./join.ts";
-export { default as handleLeave } from "./leave.ts";
-export { default as handleStartVoting } from "./start-voting.ts";
-export { default as handleStopVoting } from "./stop-voting.ts";
-export { default as handleOptionSelected } from "./option-selected.ts";
-export { default as handleModeratorChanged } from "./moderator-changed.ts";
-export { default as handleVoterKicked } from "./voter-kicked.ts";
-export { default as handleVotingDescriptionUpdated } from "./voting-desc-updated.ts";
-export { default as handleNameChanged } from "./name-changed.ts";
+import handleJoin from "./join.ts";
+import handleStartVoting from "./start-voting.ts";
+import handleStopVoting from "./stop-voting.ts";
+import handleOptionSelected from "./option-selected.ts";
+import handleModeratorChanged from "./moderator-changed.ts";
+import handleVoterKicked from "./voter-kicked.ts";
+import handleVotingDescriptionUpdated from "./voting-desc-updated.ts";
+import handleNameChanged from "./name-changed.ts";
+import {
+	validateInRoom,
+	validateModerator,
+	validateVoter,
+} from "./middlewares/mod.ts";
+import type { WebScoketMessageEvent } from "../types/socket.ts";
+import type { EventFunction } from "./types.ts";
+
+const eventHandlerMap = new Map<
+	WebScoketMessageEvent["event"],
+	Array<EventFunction<any>>
+>([
+	["Join", [handleJoin]],
+	["StartVoting", [validateModerator, handleStartVoting]],
+	["StopVoting", [validateModerator, handleStopVoting]],
+	["OptionSelected", [validateVoter, handleOptionSelected]],
+	["ModeratorChange", [validateModerator, handleModeratorChanged]],
+	["KickVoter", [validateModerator, handleVoterKicked]],
+	[
+		"UpdateVotingDescription",
+		[validateModerator, handleVotingDescriptionUpdated],
+	],
+	["ChangeName", [validateInRoom, handleNameChanged]],
+]);
+
+export default eventHandlerMap;
